+++ app/components/Table.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import $ from 'jquery'
-import Row from '../Row'
+import Row from './row'
 
 class Table extends React.Component {
 
@@ -9,27 +9,8 @@ class Table extends React.Component {
       <div>
         <table className="table table-bordered table-rounded">
           <tr><th>Self (ms)</th><th>Total (ms)</th><th>Method</th></tr>
-          {this.displayRows()}
         </table>
       </div>
-    )
-  }
-
-  displayRows() {
-    let sampleData = this.getSampleData();
-    return sampleData.forEach(sample => {
-      return (
-        <Row rowData={sample} />
-      )
-    })
-  }
-
-  getSampleData() {
-    return (
-      [ 
-        [ 1, 'A', 'B' ],
-        [ 2, 'C', 'D' ]
-      ]
     )
   }
 }
