+++ app/redux/reducer.js
@@ -1,4 +1,4 @@
-let reducer = function(state, action) {
+const reducer = function(state, action) {
   switch(action.type) {
     case 'CONNECT_TO_SERVER':
       return Object.assign({}, state, { connectionStatus: 'Connected'})
