--- app/redux/reducers/index.js
@@ -1,8 +1,0 @@
- import { combineReducers } from 'redux'
-import connectionReducer from './connectionReducer'
-
-const rootReducer = combineReducers({
-  connectionStatus: connectionReducer
-})
-
-export default rootReducer
