+++ app/redux/store.js
@@ -2,11 +2,11 @@ import { applyMiddleware, compose, createStore } from 'redux'
 import logger from 'redux-logger'
 import reducer from './reducer'
 
-let finalCreateStore = compose(
+const store = compose(
   applyMiddleware(logger())
 )(createStore)
 
 
 export default function configureStore(initialState = { connectionStatus: '' }) {
-  return finalCreateStore(reducer, initialState)
+  return store(reducer, initialState)
 }
